requirements.txt
matplotlib
numpy
python-dotenv
streamlit
seaborn
pandas
openai
langchain
langchain-experimental
langchain-anthropic
langchain-openai
langchain-community
tabulate
plotly